test: split test files into separate modules
Extract inline tests into dedicated test files for better code organization.

- Create src/lexer/tokenizer_tests.rs with 5 lexer tests
- Create src/parser/tests.rs with 5 parser tests
- Remove inline tests from src/lexer/tokenizer.rs (70 lines)
- Remove inline tests from src/parser/mod.rs (95 lines)
- Update module declarations to include test modules
